requirements.txt
fastapi>=0.100.0,<0.110.0
uvicorn[standard]>=0.23.0,<0.25.0
sqlalchemy>=2.0.0,<3.0.0
pydantic>=2.0.0,<3.0.0
python-dateutil>=2.8.0
httpx>=0.24.0,<0.26.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0  # PostgreSQL 驱动
alembic>=1.12.0  # 数据库迁移工具

